Updates the Rotor2 convention to match Rotor3
This change isn't breaking if you only interact with rotors through methods.
If you access the rotor fields you'll get a compiler error since xy became yx,
you'll need to rewrite this field in your code and negate it since its sign is
now flipped.

Please note that if you were relying on the fact that this struct is marked as
extern to share it with other languages, you won't get a compiler error and need
to catch this on your own.