Replaces arbitrary matrix transforms with affine/nonaffine transforms
It doesn't make sense to extend a homogeneous matrix to a new dimension
the way we were doing it. It's possible to do it by inserting the identity
values before z, and z becomes w, but this is not a common operation and
it's too hard to name the conversion in a way that's clear so we aren't
going to support it.